+++ scripts/cleandist.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   Distribution clean script for packaging
-##   Last updated on April 9, 2019
+##   Last updated on June 5, 2021
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2021 Paulo Moura <pmoura@logtalk.org>
@@ -45,6 +45,7 @@ chmod a-x scripts/*.js
 chmod a+x scripts/embedding/eclipse/*.sh
 chmod a+x scripts/embedding/gprolog/*.sh
 chmod a+x scripts/embedding/jiprolog/*.sh
+chmod a+x scripts/embedding/lvm/*.sh
 chmod a+x scripts/embedding/sicstus/*.sh
 chmod a+x scripts/embedding/swipl/*.sh
 chmod a+x scripts/embedding/xsb/*.sh